+++ packages/ui/src/constants/wrappedContracts.ts
@@ -28,7 +28,7 @@ const wrappedContracts: Record<number, string> = {
   3776: '0xE9CC37904875B459Fa5D0FE37680d36F1ED55e38', // astar zkEVM
   17069: '0x4200000000000000000000000000000000000006', // garnet
   690: '0x4200000000000000000000000000000000000006', // redstone
-  80085: '0x5806e416da447b267cea759358cf22cc41fae80f', // berachain testnet
+  80084: '0x7507c1dc16935b82698e4c63f2746a2fcf994df8', // berachain testnet
   666666666: '0xeb54dacb4c2ccb64f8074eceea33b5ebb38e5387', // degen
   660279: '0x3fb787101dc6be47cfe18aeee15404dcc842e6af', // xai
   1482601649: '0xab01bad2c86e24d371a13ed6367bdca819589c5d', // nebula
@@ -38,6 +38,20 @@ const wrappedContracts: Record<number, string> = {
   1329: '0xE30feDd158A2e3b13e9badaeABaFc5516e95e8C7', // sei
   1993: '0x48a9b22b80f566e88f0f1dcc90ea15a8a3bae8a4', // b3 testnet
   646: '0x48a9b22b80f566e88f0f1dcc90ea15a8a3bae8a4', // flow previewnet
+  70805: '0x48a9b22b80f566e88f0f1dcc90ea15a8a3bae8a4', // cloud
+  13746: '0x6b885d96916d18cd78e44b42c6489ca6f8794565', // game7 testnet
+  70701: '0x48a9b22b80f566e88f0f1dcc90ea15a8a3bae8a4', // boss
+  984122: '0xd5eace1274dbf70960714f513db207433615a263', // forma
+  984123: '0xd5eace1274dbf70960714f513db207433615a263', // forma sketchpad
+  8333: '0x48a9b22b80f566e88f0f1dcc90ea15a8a3bae8a4', // b3
+  33139: '0x8073b2158aa023dd7f8d4799c883b65daf6baa57', // apechain
+  33111: '0x34b097b133bf3fe3b7dad0d65e8f18dd5bf1a57b', // curtis
+  360: '0x48a9b22b80f566e88f0f1dcc90ea15a8a3bae8a4', // shape
+  11011: '0x48a9b22b80f566e88f0f1dcc90ea15a8a3bae8a4', // shape sepolia
+  11124: '0x924c922c60c5066667b63fd311880741cb4111cd', // abstract testnet
+  1946: '0x4200000000000000000000000000000000000006', // minato
+  2911: '0x2b1499d631bffb29eed7749b12cba754273d6da7', // hychain
+  29112: '0x2549584be33491340eee6762992055cda05b2581', // hychain testnet
 }
 
 export default wrappedContracts