+++ packages/gelato-adapter/CHANGELOG.md
@@ -1,5 +1,123 @@
 # @reservoir0x/gelato-adapter
 
+## 5.0.14
+
+### Patch Changes
+
+- Updated dependencies [9adc2db]
+  - @reservoir0x/reservoir-sdk@2.4.14
+
+## 5.0.13
+
+### Patch Changes
+
+- Updated dependencies [9e723b8]
+  - @reservoir0x/reservoir-sdk@2.4.13
+
+## 5.0.12
+
+### Patch Changes
+
+- Updated dependencies [55f67f6]
+  - @reservoir0x/reservoir-sdk@2.4.12
+
+## 5.0.11
+
+### Patch Changes
+
+- Updated dependencies [6118898]
+  - @reservoir0x/reservoir-sdk@2.4.11
+
+## 5.0.10
+
+### Patch Changes
+
+- Updated dependencies [ff0433f]
+  - @reservoir0x/reservoir-sdk@2.4.10
+
+## 5.0.9
+
+### Patch Changes
+
+- Updated dependencies [9a8b493]
+  - @reservoir0x/reservoir-sdk@2.4.9
+
+## 5.0.8
+
+### Patch Changes
+
+- e575a14: Sync api types
+- Updated dependencies [e575a14]
+  - @reservoir0x/reservoir-sdk@2.4.8
+
+## 5.0.7
+
+### Patch Changes
+
+- Updated dependencies [71664d9]
+  - @reservoir0x/reservoir-sdk@2.4.7
+
+## 5.0.6
+
+### Patch Changes
+
+- Updated dependencies [3aae35c]
+  - @reservoir0x/reservoir-sdk@2.4.6
+
+## 5.0.5
+
+### Patch Changes
+
+- Updated dependencies [c3a53b6]
+  - @reservoir0x/reservoir-sdk@2.4.5
+
+## 5.0.4
+
+### Patch Changes
+
+- Updated dependencies [6da031f]
+  - @reservoir0x/reservoir-sdk@2.4.4
+
+## 5.0.3
+
+### Patch Changes
+
+- Updated dependencies [d08dccf]
+- Updated dependencies [ee9b0e8]
+  - @reservoir0x/reservoir-sdk@2.4.3
+
+## 5.0.2
+
+### Patch Changes
+
+- Updated dependencies [e46059b]
+  - @reservoir0x/reservoir-sdk@2.4.2
+
+## 5.0.1
+
+### Patch Changes
+
+- Updated dependencies [49fc944]
+  - @reservoir0x/reservoir-sdk@2.4.1
+
+## 5.0.0
+
+### Minor Changes
+
+- b9063e3: Upgrade viem and wagmi
+
+### Patch Changes
+
+- Updated dependencies [b9063e3]
+  - @reservoir0x/reservoir-sdk@2.4.0
+
+## 4.0.1
+
+### Patch Changes
+
+- Updated dependencies [9686ca3]
+  - @reservoir0x/reservoir-sdk@2.3.1
+
 ## 4.0.0
 
 ### Minor Changes